C++ Warning fixes, werror on ci, C++17 compliance (#2957)
### What

* Fixes #2903 
* Part of #2919
* currate warning list for gcc/clang
* enable warnings as errors on CI, for simplicity it's always active on
roundtrip tests
* enable standard set of warning settings on all C++ targets
* fix warnings in codegen'ed code and manually written code
* fix accidental C++20 (should have been caught in theory by warnings
but it didn't trigger for unknown reason)

### Checklist
* [x] I have read and agree to [Contributor
Guide](https://github.com/rerun-io/rerun/blob/main/CONTRIBUTING.md) and
the [Code of
Conduct](https://github.com/rerun-io/rerun/blob/main/CODE_OF_CONDUCT.md)
* [x] I've included a screenshot or gif (if applicable)
* [x] I have tested [demo.rerun.io](https://demo.rerun.io/pr/2957) (if
applicable)

- [PR Build Summary](https://build.rerun.io/pr/2957)
- [Docs
preview](https://rerun.io/preview/pr%3Aandreas%2Fcpp%2Fwarning-fixes-and-ci-werror/docs)
- [Examples
preview](https://rerun.io/preview/pr%3Aandreas%2Fcpp%2Fwarning-fixes-and-ci-werror/examples)

+++ rerun_cpp/src/rerun/datatypes/class_id.cpp
@@ -33,7 +33,9 @@ namespace rerun {
             }
 
             static_assert(sizeof(*elements) == sizeof(elements->id));
-            ARROW_RETURN_NOT_OK(builder->AppendValues(&elements->id, num_elements));
+            ARROW_RETURN_NOT_OK(
+                builder->AppendValues(&elements->id, static_cast<int64_t>(num_elements))
+            );
 
             return arrow::Status::OK();
         }
